requirements.txt
faiss-cpu
sentence-transformers
PyPDF2
requests
python-dotenv
langdetect
sambanova
numpy
langchain_sambanova
types-requests>=2.28.0
pytest>=7.0.0
mypy>=0.950
black>=22.0.0
language-tool-python
fastapi 
uvicorn